requirements.txt
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.5.0
Django==1.10.5
django-filter==1.0.1
django-sendfile==0.3.11
django-unixtimestampfield==0.3.9
djangorestframework==3.7.3
graphene==1.1.3
graphene-django==1.2.1
graphql-core==1.0.1
graphql-relay==0.4.5
iso8601==0.1.11
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
oauthlib==2.0.1
olefile==0.44
Pillow==4.0.0
promise==1.0.1
psycopg2
PyJWT==1.4.2
python3-openid==3.0.10
PyYAML==3.12
requests==2.13.0
requests-oauthlib==0.7.0
singledispatch==3.4.0.3
six==1.10.0
social-auth-app-django==1.1.0
social-auth-core==1.2.0
typing==3.5.3.0
uritemplate==3.0.0
python-memcached==1.59